Fixed error
I downloaded the SDK from scratch and put our code in the TeamCode folder. I know that git is made for doing these kinds of things automatically, but it shouldn't be necessary on a regular basis and this was easier for me. The code now builds properly on my computer.